Separated parser into another file
Separated feature1.js implementation into two parts:
- Created a dedicated function for parsing file to send to data-confirm
  page for spreadsheet
- Implemented try catch errors in all 3 features
