+++ src/backend/feature1.js
@@ -6,69 +6,15 @@ const Papa = require('papaparse'); // For CSV parsing
 const mammoth = require('mammoth'); // For Docx parsing
 const prompts = require('./prompts/deepseekPrompts');
 
-async function parseFileAndSendToDeepSeek(file, query){
-    let filePath = null;
-    try{
-        let data = [];
-        filePath = file.path;
-
-        switch (file.mimetype) {
-            //parse csv
-            case 'text/csv': {
-                const csvContent = fs.readFileSync(filePath, 'utf8');
-                const parsedCsv = Papa.parse(csvContent, { header: true });
-                data = parsedCsv.data;
-                break;
-            }
-            
-            //parse excel
-            case 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet': {
-                const workbook = XLSX.readFile(file.path);
-                const sheet = workbook.Sheets[workbook.SheetNames[0]];
-                data = XLSX.utils.sheet_to_json(sheet);
-                break;
-            }
-
-            //parse pdf
-            case 'application/pdf': {
-                const pdfContent = fs.readFileSync(file.path);
-                const pdfText = await pdfParse(pdfContent);
-                data = pdfText.text.split('\n');
-                break;
-            }
-
-            //parse docx
-            case 'application/vnd.openxmlformats-officedocument.wordprocessingml.document': {
-                const docxBuffer = fs.readFileSync(file.path);
-                const docxText = await mammoth.extractRawText({ buffer: docxBuffer });
-                data = docxText.value.split('\n');
-                break;
-            }
-
-            //parse plain text
-            case 'text/plain': {
-                const txtContent = fs.readFileSync(file.path, 'utf8');
-                data = txtContent.split('\n');
-                break;
-            }
-
-            //if non-matched
-            default:
-                throw new Error('Unsupported file type.');
-        }
-
-        //send to deepseek api
-        const prompt = prompts.feature1(query, data);
+async function convertToChartConfig(query, data) {
+    const prompt = prompts.feature1(query, data);
+    try {
         const result = await queryDeepSeekV3(prompt);
-        return {analysis: JSON.parse(result)};
+        return JSON.parse(result);
     } catch (error) {
-        console.error('Error processing file:', error);
+        console.error('Error converting to chart config:', error);
         throw error;
-    } finally {
-        if (filePath && fs.existsSync(filePath)) {
-            fs.unlinkSync(filePath);
-        }
     }
 }
 
-module.exports = {parseFileAndSendToDeepSeek};
+module.exports = {convertToChartConfig};
