+++ src/backend/feature2.js
@@ -4,8 +4,13 @@ const prompts = require('./prompts/deepseekPrompts.js');
 
 async function getGraphRecommendation(data){
     const prompt = prompts.feature2("", data);
-    const result = await queryDeepSeekV3(prompt);
-    return result;
+    try {
+        const result = await queryDeepSeekV3(prompt);
+        return result;
+    } catch (error) {
+        console.error('Error getting graph recommendation:', error);
+        throw error;
+    }
 }
 
 module.exports = {getGraphRecommendation};
