+++ src/backend/feature3.js
@@ -4,8 +4,13 @@ const prompts = require('./prompts/deepseekPrompts.js');
 
 async function getSummary(data){
     const prompt = prompts.feature3("", data);
-    const result = await queryDeepSeekV3(prompt);
-    return result;
+    try {
+        const result = await queryDeepSeekV3(prompt);
+        return result;
+    } catch (error) {
+        console.error('Error getting summary:', error);
+        throw error;
+    }
 }
 
 module.exports = {getSummary};
